--- app/src/main/java/com/lambdaschool/sprint2_challenge/MainActivity.kt
@@ -1,12 +1,0 @@
-package com.lambdaschool.sprint2_challenge
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
