+++ app/src/main/java/com/lambdaschool/sprint2_challenge/model/ShoppingItem.kt
@@ -1,0 +1,8 @@
+package com.lambdaschool.sprint2_challenge.model
+
+class ShoppingItem(itemName:String,itemImageID:Int,itemIndex:Int){
+    val itemName:String=itemName
+    val itemImageId:Int=itemImageID
+    val itemIndex:Int=itemIndex
+
+}
