+++ app/src/main/java/com/lambdaschool/sprint2_challenge/ui/ShoppingListAdapter.kt
@@ -1,0 +1,43 @@
+package com.lambdaschool.sprint2_challenge.ui
+import android.support.v7.content.res.AppCompatResources.getDrawable
+import android.support.v7.widget.RecyclerView
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import com.lambdaschool.sprint2_challenge.R
+import com.lambdaschool.sprint2_challenge.model.ShoppingItem
+import java.util.*
+
+class ShoppingListAdapter (val data: ArrayList<ShoppingItem>) :
+        RecyclerView.Adapter<ShoppingListAdapter.ViewHolder>() {
+
+    class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+        val itemImage: ImageView = view.findViewById(R.id.item_image)
+        val itemName: TextView = view.findViewById(R.id.item_name)
+
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val viewGroup = LayoutInflater.from(parent.context).inflate(R.layout.shopping_item_layout, parent, false)
+        return ViewHolder(viewGroup)
+    }
+
+    override fun getItemCount(): Int {
+        return data.size
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+       holder.itemImage.setImageDrawable(getDrawable(holder.itemImage.context,data[position].itemImageId))
+        holder.itemName.text=  data[position].itemName
+
+        holder.itemView.setOnClickListener {
+
+        }
+
+
+
+    }
+}
+
